Added effects table instead of json field

+++ database/migrations/2019_11_15_151747_create_pages_table.php
@@ -15,9 +15,10 @@ class CreatePagesTable extends Migration
     {
         Schema::create('pages', function (Blueprint $table) {
             $table->uuid('uuid')->primary();
+
             $table->unsignedInteger('number');  //FIXME: can be deleted, useless
 
-            $table->unsignedInteger('story_id');
+            $table->unsignedBigInteger('story_id');
             $table->foreign('story_id')->references('id')->on('stories');
 
             $table->boolean('is_first')->default(false);
