+++ database/migrations/2019_12_09_142743_create_actions_table.php
@@ -15,11 +15,14 @@ class CreateActionsTable extends Migration
     {
         Schema::create('actions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('item_id');
+
+            $table->unsignedBigInteger('item_id');
             $table->foreign('item_id')->references('id')->on('items');
+
             $table->string('verb');
             $table->integer('quantity')->default(1);
             $table->integer('price')->nullable();
+
             $table->uuid('page_uuid');
             $table->foreign('page_uuid')->references('uuid')->on('pages');
         });
